Update location validator
The API we're currently using (Mapbox API) allows us to specify the type of the location we're querying. This allows for better filtering of results and since the results more refined, there's a less likely chance that location validation may fail. This, however, does not mean success is always guaranteed (or in other words, fault-free validation) as we rely solely on the API.