Add global collapsible alert
This collapsible alert is ideal for messages. Forms inheriting from the default template just need to add an 'errors' property (which is a collection of strings) on the model and the alert should pop up on the page if and only if the 'errors' property is not empty.

I also made changes to the 'content-body' fragment of the 'default'
template. It is no longer enclosed in a <div class="row">...</div> element
and inheriting classes shold therefore declare this on their own. Doing so
affected all the files inheriting from 'default.html'.